requirements.txt
absl-py==2.1.0
astunparse==1.6.3
audioread==3.0.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cycler==0.11.0
decorator==5.1.1
docopt==0.6.2
ffmpeg==1.4
Flask==2.2.5
Flask-Cors==5.0.0
flatbuffers==23.5.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.28.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.0
gunicorn==23.0.0
h5py==3.8.0
idna==3.6
importlib-metadata==6.7.0
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
Js2Py==0.74
jsonpickle==3.0.3
keras==2.11.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.39.1
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
more-itertools==9.1.0
msgpack==1.0.5
music21==7.3.3
numba==0.56.4
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pipwin==0.5.2
platformdirs==4.0.0
pooch==1.8.1
prompt-toolkit==3.0.43
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydub==0.25.1
Pygments==2.17.2
pyjsparser==2.7.1
PyLyrics==1.1.0
pyparsing==3.1.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.7
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.16.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2024.1
tzlocal==5.1
urllib3==2.0.7
wcwidth==0.2.13
webcolors==1.13
Werkzeug==2.2.3
wget==3.2
wrapt==1.16.0
zipp==3.15.0
tqdm==4.67.1
gdown==4.7.3
filelock==3.12.2
PySocks==1.7.1
